+++ crates/teloxide-core/CHANGELOG.md
@@ -117,10 +117,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [pr850]: https://github.com/teloxide/teloxide/pull/850
 
 ### Removed
+
 - Remove `can_send_media_messages` from `ChatPermissions` ([#954][pr954])
 - Remove `can_send_media_messages` field from `Restricted` ([#954][pr954])
+- Previously deprecated items ([#1013][pr1013])
+  - `AutoSend` bot adaptor
 
 [pr954]: https://github.com/teloxide/teloxide/pull/954
+[pr1013]: https://github.com/teloxide/teloxide/pull/1013
 
 ### Fixed